refactor: add group.shield-rules file
