+++ src/auth/shield/shield.permissions.ts
@@ -3,24 +3,14 @@ import { FORBIDDEN } from "../../common/common.constants";
 import {
   canApproveGroupMemberRequests,
   canCreateGroupEvents,
-  canCreateServerInvites,
   canDeleteGroup,
-  canManageEvents,
   canManageGroupEvents,
   canManageGroupPosts,
   canManageGroupRoles,
   canManageGroupSettings,
-  canManagePosts,
-  canManageServerInvites,
-  canManageServerRoles,
-  canRemoveMembers,
   canUpdateGroup,
-  hasValidRefreshToken,
-  isAuthenticated,
   isGroupMember,
-  isOwnPost,
   isProposalGroupJoinedByMe,
-  isPublicEvent,
   isPublicGroup,
   isPublicGroupEventImage,
   isPublicGroupImage,
@@ -28,13 +18,25 @@ import {
   isPublicGroupPostImage,
   isPublicGroupProposal,
   isPublicGroupProposalAction,
+  isPublicGroupProposalImage,
   isPublicGroupRole,
+  isPublicGroupVote,
+} from "./rules/group.shield-rules";
+import {
+  canCreateServerInvites,
+  canManageEvents,
+  canManagePosts,
+  canManageServerInvites,
+  canManageServerRoles,
+  canRemoveMembers,
+  hasValidRefreshToken,
+  isAuthenticated,
+  isOwnPost,
+  isPublicEvent,
   isUserAvatarInPublicFeed,
   isUserAvatarInPublicPost,
   isUserInPublicFeed,
   isUserInPublicPost,
-  isPublicGroupVote,
-  isPublicGroupProposalImage,
 } from "./shield.rules";
 
 export const shieldPermissions = shield(
